backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Azure SDK packages
azure-storage-blob==12.19.0
azure-identity==1.15.0
azure-ai-ml==1.12.1
azure-keyvault-secrets==4.7.0
azure-monitor-opentelemetry==1.2.0

# Database and ORM
sqlalchemy==2.0.23
alembic==1.13.1

# ML and Computer Vision (updated versions)
torch==2.6.0
torchvision==0.21.0
transformers==4.36.2
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.4
scikit-learn==1.3.2

# Image processing and OCR
pytesseract==0.3.10
easyocr==1.7.0

# Data validation and serialization
pydantic==2.5.2
pydantic-settings==2.1.0

# HTTP client and utilities
httpx==0.25.2
python-dotenv==1.0.0
loguru==0.7.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1

# MLflow for experiment tracking
mlflow==2.8.1
azureml-mlflow==1.54.0 